requirements.txt
uvicorn~=0.29.0
fastapi~=0.109.2
psycopg2-binary
tzdata
pydantic-settings
python-dotenv~=1.0.1
hypercorn
pyyaml~=6.0.1
PyJWT~=2.8.0
flask~=3.0.2
hpack~=4.0.0
hyperframe~=6.0.1
h2~=4.1.0
h11~=0.14.0
pip~=24.0
wheel~=0.41.2
Jinja2~=3.1.3
redis~=5.0.3
sniffio~=1.3.1
click~=8.1.7
starlette~=0.36.3
pydantic~=2.6.4
anyio~=4.3.0
wsproto~=1.2.0
priority~=2.0.0
itsdangerous~=2.1.2
setuptools~=68.2.0
